Fix Bank search area
The library is trying to search for items in the AREA_MS section of the screen (which is the playable area). This messes up if you use tabs and have an item in the first slot of a tab that needs to be found. AREA_BS only searches the actual clickable area.